+++ src/components/CardListContenido.jsx
@@ -1,21 +1,21 @@
-import React, { useEffect, useRef } from 'react'
+// import React, { useEffect, useRef } from 'react'
 import { v4 as uuidv4 } from 'uuid'
 import CardContenido from './CardContenido.jsx'
-import ScrollReveal from 'scrollreveal'
+// import ScrollReveal from 'scrollreveal'
 
 const CardListContenido = () => {
-  const revealRef = useRef()
+  // const revealRef = useRef()
 
-  useEffect(() => {
-    const sr = ScrollReveal({
-      duration: 1500,
-      delay: 300,
-      distance: '50px',
-      easing: 'ease-in-out'
-    })
+  // useEffect(() => {
+  //   const sr = ScrollReveal({
+  //     duration: 1500,
+  //     delay: 300,
+  //     distance: '50px',
+  //     easing: 'ease-in-out'
+  //   })
 
-    sr.reveal('.card-contenido', { interval: 100 })
-  }, [])
+  //   sr.reveal('.card-contenido', { interval: 100 })
+  // }, [])
 
   const imgBg = 'https://res.cloudinary.com/dpiwmbsog/image/upload/v1686860782/consultora/Group_1_h9hlha.svg'
 
@@ -32,7 +32,7 @@ const CardListContenido = () => {
       {
       contenido.map(item => {
         return (
-          <CardContenido referencia={revealRef} key={uuidv4()} item={item} imgBg={imgBg} />
+          <CardContenido key={uuidv4()} item={item} imgBg={imgBg} />
         )
       })
       }
