+++ src/components/CardPrecentacion.jsx
@@ -1,10 +1,24 @@
-import React from 'react'
+// import React, { useEffect } from 'react'
 import { FaRegSmileBeam, FaHandshake } from 'react-icons/fa'
+// import ScrollReveal from 'scrollreveal'
 
 const CardPrecentacion = () => {
+  // const revealRef = useRef()
+
+  // useEffect(() => {
+  //   const sr = ScrollReveal({
+  //     duration: 1500,
+  //     delay: 300,
+  //     distance: '50px',
+  //     easing: 'ease-in-out'
+  //   })
+
+  //   sr.reveal('.card-precentacion', { interval: 100 })
+  // }, [])
+
   return (
     <section className='cardPrecentacion w-full max-w-[1650px] h-auto bg-whiteWi flex flex-col justify-start items-center gap-4 p-5  sm:flex-row sm:justify-evenly sm:items-center md:gap-4 md:p-6 mt-10 lg:mt-20'>
-      <section className='w-full h-[290px] sm:w-1/2 sm:h-[290px] md:w-1/2 md:h-[270px] lg:w-1/2 lg:h-[280px] xl:w-1/2 xl:h-[290px] flex flex-col justify-center items-center p-5 overflow-hidden gap-2 lg:gap-3'>
+      <section className='card-scroll w-full h-[290px] sm:w-1/2 sm:h-[290px] md:w-1/2 md:h-[270px] lg:w-1/2 lg:h-[280px] xl:w-1/2 xl:h-[290px] flex flex-col justify-center items-center p-5 overflow-hidden gap-2 lg:gap-3'>
         <div className='w-[48px] h-[48px] rounded-full bg-[#8ad2da] text-purpleBal flex justify-center items-center text-3xl'>
           <FaHandshake />
         </div>
@@ -12,7 +26,7 @@ const CardPrecentacion = () => {
         <p className='w-full font-parrafo font-[300] text-sm sm:text-md md:text-md lg:text-lg xl:text-lg text-center'>Perfeccionar la experiencia del cliente, identificar oportunidades de mejora, implementar soluciones efectivas, aumentar la satisfacción y lealtad del cliente y mejorar la reputación de la marca.</p>
       </section>
 
-      <section className='w-full h-[290px] sm:w-1/2 sm:h-[290px] md:w-1/2 md:h-[270px] lg:w-1/2 lg:h-[280px] xl:w-1/2 xl:h-[290px] flex flex-col justify-center items-center p-5 overflow-hidden gap-2 lg:gap-3'>
+      <section className='card-scroll w-full h-[290px] sm:w-1/2 sm:h-[290px] md:w-1/2 md:h-[270px] lg:w-1/2 lg:h-[280px] xl:w-1/2 xl:h-[290px] flex flex-col justify-center items-center p-5 overflow-hidden gap-2 lg:gap-3'>
         <div className='w-[48px] h-[48px] rounded-full bg-[#8ad2da] text-purpleBal flex justify-center items-center text-3xl'>
           <FaRegSmileBeam />
         </div>
