+++ src/components/ContactosClient.jsx
@@ -22,7 +22,7 @@ const ContactosClient = () => {
   }
 
   return (
-    <section className='contenidoContacto w-full lg:w-1/2 flex flex-col gap-6 p-6'>
+    <section className='contenidoContacto card-scroll w-full lg:w-1/2 flex flex-col gap-6 p-6'>
       <h3 className='font-parrafo font-[700] text-lg uppercase'>Contactanos</h3>
       <h2 className='font-titulo font-[700] text-5xl'>Conectemos hoy</h2>
       <p className='font-parrafo font-[400] text-lg text-[#333]'>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Rem voluptas maiores harum iusto enim aperiam, necessitatibus nisi qui velit fuga veniam est? Architecto, beatae quis perspiciatis magni blanditiis nisi eveniet!</p>
