+++ src/context/GeneralContext.jsx
@@ -1,4 +1,5 @@
-import { createContext, useContext, useRef, useState } from 'react'
+import { createContext, useContext, useEffect, useRef, useState } from 'react'
+import ScrollReveal from 'scrollreveal'
 
 // se crea el contexto
 export const GeneralContext = createContext([])
@@ -12,6 +13,17 @@ export const GeneralContextProvider = ({ children }) => {
   const contactoRef = useRef(null)
   const serviciosRef = useRef(null)
   const sobremiRef = useRef(null)
+
+  useEffect(() => {
+    const sr = ScrollReveal({
+      duration: 1500,
+      delay: 300,
+      distance: '50px',
+      easing: 'ease-in-out'
+    })
+
+    sr.reveal('.card-scroll', { interval: 100 })
+  }, [])
 
   return (
     <GeneralContext.Provider
